+++ src/main.cpp
@@ -842,6 +842,10 @@ int option_handler(void* user, const char* section, const char* name, const char
     {
         opt_list_parse(cf->facility_field_defense_bonus, buf, 4, 0);
     }
+    else if (MATCH("air_superiority_not_required_to_attack_needlejet"))
+    {
+        cf->air_superiority_not_required_to_attack_needlejet = (atoi(value) == 0 ? false : true);
+    }
 	else if (MATCH("ai_useWTPAlgorithms"))
     {
         cf->ai_useWTPAlgorithms = (atoi(value) == 0 ? false : true);