+++ src/main.h
@@ -19,7 +19,7 @@
 #pragma once
 
 #ifdef BUILD_REL
-    #define MOD_VERSION "Thinker Mod v5.1 - The Will to Power mod v399"
+    #define MOD_VERSION "Thinker Mod v5.1 - The Will to Power mod v400"
 #else
     #define MOD_VERSION "Thinker Mod develop build"
 #endif
@@ -449,6 +449,7 @@ struct Config {
     bool scorched_earth = false;
     bool destroy_captured_base_defense = false;
     int facility_field_defense_bonus[4] = {0,0,0,0};
+    bool air_superiority_not_required_to_attack_needlejet = false;
     // AI configurations
     bool ai_useWTPAlgorithms;
     bool wtp_enabled_factions[MaxPlayerNum];